Improved receipt functionality
After taking a picture of a receipt, clicking the thumbnail still
brings you to the camera to take a new picture. That's not a bug;
I just haven't implemented the "view receipt" functionality yet.
That should just be a question of making the thumbnail show a bigger version
of the receipt when clicked and adding a separate "take picture"
button, though. The functionality should be there though (although
I'll have to do more extensive testing with Junit).