Calendar bug fix
When making and updating a claim, when the start date was
the same as the end date,
choosing the end date before choosing the start date gave a
"The end date must be after the start date" error because
the time was saved as well as the date. Fixed this by
setting the time to 0.